fix(build): make the Linux GCC build warning-free and green
Bring the Linux build to the same standard as the canonical Clang one: a clean
GCC 15 compile under warnings-as-errors and a fully passing test suite. Close
the toolchain, linker, code generation, and HiDPI defects that only surface
where GCC, glibc, and Wayland differ from the macOS toolchain, fixing each at
its cause so no platform is special-cased into correctness.

Toolchain and warning surface

- scope -Wno-overloaded-virtual and -fvisibility-inlines-hidden to C++
  translation units; GCC diagnoses both when they reach a C source, and under
  -Werror the option itself becomes the error
- select -Wshadow=local on GCC, suppress the plural-spelled unused-local-typedef
  warning, and route indirect callback casts through RT_FN_PTR_CAST
- guard __builtin_popcountll behind hardware population-count support so the
  native linker is never asked for libgcc archive members
- assemble through cc, not a cross-target clang invocation, when the codegen
  pipeline targets its own host; a distribution whose default toolchain is GCC
  could not use system-assembler mode at all
- report a skip rather than a failure from the sanitizer and coverage
  self-tests when the Clang toolchain they verify is absent

Native linker

- resolve imported data objects through copy relocations with sized, defined
  .dynsym entries; representing them as jump thunks made every reference to
  stdout load instruction bytes as a FILE pointer, so no native binary that
  printed could run
- give each such import a separate GOT word holding its address, so a
  translation unit reaching it through GOTPCREL does not dereference one level
  too far
- emit .gnu.version and .gnu.version_r from the providing libraries' default
  versions; an unversioned reference does not reliably select a library's
  current definition, and pthread_cond_init, memcpy, pow, log, and
  clock_gettime were all binding to compatibility implementations
- stop string dedup from merging a section symbol as the string that starts its
  section: a mergeable string section is addressed as "section + offset", so
  merging it redirected every one of those offsets into another object's copy
  of that first string

Code generation

- keep a pre-allocation copy whose destination is read by a successor block;
  the forwarding scan stops at the terminator, so proving single use inside one
  block deleted the definition out from under the consumer and left it reading
  whatever the allocator later parked there

Graphics and HiDPI

- record whether a coordinate scale was chosen explicitly; the GUI layer's
  untouched 1.0 default is otherwise indistinguishable from a deliberate 1.0,
  so the first asynchronous scale report flipped every window into
  coordinate-scaled mode and public sizes were divided by the scale twice
- re-derive the framebuffer from the requested logical size once a backend that
  learns its HiDPI factor only after the surface exists has published it, so
  the window is settled before the first frame instead of resizing under it
- treat framebuffer name zero as the default framebuffer rather than a member
  of the target set being replaced, which had the restore path replay a
  default-framebuffer draw buffer against a new FBO and leave
  GL_INVALID_OPERATION queued for the next caller
- implement set_capture_after_present for OpenGL, whose readback source is the
  back buffer that presenting leaves undefined

Coverage and probes

- add regressions for cross-block copy forwarding on both backends, for symbol
  version requirements, and for section-symbol string merging
- fall back to a realtime-clock condition variable when an implementation
  rejects the monotonic attribute at construction
- pump the hidden-outline fold check until it lands, as its visible counterpart
  already does, so a correctness assertion stops doubling as a host-speed
  budget; let the software render baseline match within one pixel, since
  transcendental results differ between C libraries and move a silhouette
- declare SetCaptureAfterPresent in the shadow probe, which reads pixels back
  after presenting

Validated with a clean scripts/build_zanna_linux.sh run: no compiler warnings,
1874 of 1874 tests passing, and the platform policy lint, runtime surface
audit, and cross-platform smoke stages all clean.

+++ src/runtime/game/rt_scene_editor.cpp
@@ -5005,7 +5005,7 @@ int64_t rt_game_scene_tile_collision(void *scene, int64_t tile) {
     SCENE_TRY {
         const auto &collision = requireScene(scene)->state->behavior.collision;
         auto it = collision.find(tile);
-        return it == collision.end() ? RT_TILE_COLLISION_NONE : it->second;
+        return it == collision.end() ? static_cast<int64_t>(RT_TILE_COLLISION_NONE) : it->second;
     }
     SCENE_CATCH(RT_TILE_COLLISION_NONE)
 }
